+++ R/03-modelResults3DKernel.R
@@ -347,7 +347,7 @@ modelResults3DKernelUI <- function(id, title = ""){
                             selected = "RdYlGn"),
                 ns = ns),
               selectInput(inputId = ns("pointShape"), label = "Shape of location marks",
-                          choices = 0:25, selected = 4),
+                          choices = pchChoices(), selected = 4),
               ns = ns),
             checkboxInput(inputId = ns("cluster"),
                           label = "Show Clustering",