+++ NEWS.md
@@ -1,6 +1,6 @@
 # MpiIsoApp development version
 
-## Version 23.04.2
+## Version 23.04.3
 
 ### Updates
   - removing the calculation of clusters for unfiltered data
@@ -8,6 +8,16 @@
   - renaming "Number of clusters (optional)" to "Possible range for clusters"
   - increasing the maximum number of clusters from 20 to 50 for mclust
   - add option to show spatial centroids
+
+## Version 23.04.2
+
+### New Features
+- _Interactive map_: 
+  - option to set the symbol dependent on values of a character column (#13)
+  - show/hide legends for size (if a variable is used) and symbols 
+
+### Updates
+- optimized behavior of updating data and show/hide legends on the interactive map
 
 ## Version 23.04.1